Outline, segfaults when compiling
