+++ pfb/wavelets/wavelets.py
@@ -1,5 +1,4 @@
 from collections import namedtuple
-from enum import Enum
 
 import numba
 import numba.core.types as nbtypes
@@ -8,6 +7,8 @@ from numba.extending import register_jitable, overload
 import numpy as np
 
 from pfb.wavelets.coefficients import coefficients
+from pfb.wavelets.convolution import downsampling_convolution
+from pfb.wavelets.modes import Modes, promote_mode
 from pfb.wavelets.utils import slice_axis, force_type_contiguity
 
 
@@ -30,67 +31,7 @@ DiscreteWavelet = namedtuple("DiscreteWavelet",
                                 "vanishing_moments_psi",
                                 "vanishing_moments_phi"))
 
-_VALID_MODES = ["zeropad", "symmetric", "constant_edge",
-                "smooth", "periodic", "periodisation",
-                "reflect", "asymmetric", "antireflect"]
 
-
-class Modes(Enum):
-    zeropad = 0
-    symmetric = 1
-    constant_edge = 2
-    smooth = 3
-    periodic = 4
-    periodisation = 4
-    reflect = 5
-    asymmetric = 6
-    antireflect = 7
-
-
-def mode_str_to_enum(mode_str):
-    pass
-
-
-@overload(mode_str_to_enum)
-def mode_str_to_enum_impl(mode_str):
-    if isinstance(mode_str, nbtypes.UnicodeType):
-        # Modes.zeropad.name doesn't work in the jitted code
-        # so expand it all.
-        zeropad_name = Modes.zeropad.name
-        symmetric_name = Modes.symmetric.name
-        constant_edge_name = Modes.constant_edge.name
-        smooth_name = Modes.smooth.name
-        periodic_name = Modes.periodic.name
-        periodisation_name = Modes.periodisation.name
-        reflect_name = Modes.reflect.name
-        asymmetric_name = Modes.asymmetric.name
-        antireflect_name = Modes.antireflect.name
-
-        def impl(mode_str):
-            mode_str = mode_str.lower()
-
-            if mode_str == zeropad_name:
-                return Modes.zeropad
-            elif mode_str == symmetric_name:
-                return Modes.symmetric
-            elif mode_str == constant_edge_name:
-                return Modes.constant_edge
-            elif mode_str == smooth_name:
-                return Modes.smooth
-            elif mode_str == periodic_name:
-                return Modes.periodic
-            elif mode_str == periodisation_name:
-                return Modes.periodisation
-            elif mode_str == reflect_name:
-                return Modes.reflect
-            elif mode_str == asymmetric_name:
-                return Modes.asymmetric
-            elif mode_str == antireflect_name:
-                return Modes.antireflect
-            else:
-                raise ValueError("Unknown mode string")
-
-        return impl
 
 
 @register_jitable
@@ -114,7 +55,7 @@ def dwt_buffer_length(input_length, filter_length, mode):
     if input_length < 1 or filter_length < 1:
         return 0
 
-    if mode == "periodisation":
+    if mode is Modes.periodisation:
         return (input_length // 2) + (1 if input_length % 2 else 0)
     else:
         return (input_length + filter_length - 1) // 2
@@ -225,55 +166,6 @@ def promote_axis(axis, ndim):
     return impl
 
 
-@numba.generated_jit(nopython=True, nogil=True, cache=True)
-def promote_mode(mode, naxis):
-    if not isinstance(naxis, nbtypes.Integer):
-        raise TypeError("naxis must be an integer")
-
-    if isinstance(mode, nbtypes.misc.UnicodeType):
-        def impl(mode, naxis):
-            return numba.typed.List([mode_str_to_enum(mode) for _ in range(naxis)])
-
-    elif ((isinstance(mode, nbtypes.containers.List) or
-          isinstance(mode, nbtypes.containers.UniTuple)) and
-            isinstance(mode.dtype, nbtypes.UnicodeType)):
-        def impl(mode, naxis):
-            if len(mode) != naxis:
-                raise ValueError("len(mode) != len(axis)")
-
-            return numba.typed.List([mode_str_to_enum(m) for m in mode])
-    else:
-        raise TypeError("mode must be a string, "
-                        "a list of strings "
-                        "or a tuple of strings. "
-                        "Got %s." % mode)
-
-    return impl
-
-
-@numba.generated_jit(nopython=True, nogil=True, cache=True)
-def downsampling_convolution(input, N, filter,
-                             output, step, mode):
-    def impl(input, N, filter, F, output,
-             step, mode):
-        i = step - 1
-        o = 0
-
-        if mode == "smooth" and N < 2:
-            mode = "constant_edge"
-
-        while i < F and i < N:
-            sum = input.dtype.type(0)
-
-            for j in range(i):
-                sum += filter[j] * input[i-j]
-
-            i += step
-            o += 1
-
-    return impl
-
-
 @numba.generated_jit(nopython=True, nogil=True)
 def dwt_axis(data, wavelet, mode, axis):
     def impl(data, wavelet, mode, axis):
@@ -293,7 +185,8 @@ def dwt_axis(data, wavelet, mode, axis):
         # Iterate over all points except along the slicing axis
         for idx in np.ndindex(*tuple_setitem(data.shape, axis, 1)):
             initial_in_row = slice_axis(data, idx, axis)
-            initial_out_row = slice_axis(ca, idx, axis)
+            initial_ca_row = slice_axis(ca, idx, axis)
+            initial_cd_row = slice_axis(cd, idx, axis)
 
             # The numba array type returned by slice_axis assumes
             # non-contiguity in the general case.
@@ -306,10 +199,24 @@ def dwt_axis(data, wavelet, mode, axis):
             else:
                 in_row = initial_in_row.copy()
 
-            if initial_out_row.flags.c_contiguous:
-                out_row = force_type_contiguity(initial_out_row)
+            if initial_ca_row.flags.c_contiguous:
+                ca_row = force_type_contiguity(initial_ca_row)
             else:
-                out_row = initial_out_row.copy()
+                ca_row = initial_ca_row.copy()
+
+            if initial_cd_row.flags.c_contiguous:
+                cd_row = force_type_contiguity(initial_cd_row)
+            else:
+                cd_row = initial_cd_row.copy()
+
+            # downsampling_convolution(in_row, ca_row, wavelet.dec_lo, mode)
+            # downsampling_convolution(in_row, cd_row, wavelet.dec_hi, mode)
+
+            if not initial_ca_row.flags.c_contiguous:
+                initial_ca_row[:] = ca_row[:]
+
+            if not initial_cd_row.flags.c_contiguous:
+                initial_cd_row[:] = cd_row[:]
 
         return ca, cd
